requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.2
pdfminer.six>=20221105
python-docx>=0.8.11
python-pptx>=0.6.21
loguru>=0.5.3
httpx>=0.23.0